The MachineModel now computes all sizes as BigIntegers
to be able to handle type sizes that are larger than 2^63.

This fixes #426.


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@29904 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/value/AbstractExpressionValueVisitor.java
@@ -1407,7 +1407,7 @@ public abstract class AbstractExpressionValueVisitor
 
     switch (idOperator) {
     case SIZEOF:
-      int size = machineModel.getSizeof(innerType);
+        BigInteger size = machineModel.getSizeof(innerType);
       return new NumericValue(size);
 
     default: // TODO support more operators
